rename drop.ts to use-drop.ts
